+++ cipher/viper_test.go
@@ -56,3 +56,46 @@ func TestNOOPViper(t *testing.T) {
 	assert.NoError(err)
 	assert.Equal([]byte(msg), data)
 }
+
+func TestBoxBothSides(t *testing.T) {
+	assert := assert.New(t)
+
+	vSend := viper.New()
+	path, err := os.Getwd()
+	assert.NoError(err)
+	vSend.AddConfigPath(path)
+	vSend.SetConfigName("boxSender")
+	if err := vSend.ReadInConfig(); err != nil {
+		t.Fatalf("%s\n", err)
+	}
+
+	options, err := FromViper(vSend)
+	assert.NoError(err)
+
+	encrypter, err := options.LoadEncrypt()
+	assert.NoError(err)
+
+	vRec := viper.New()
+	assert.NoError(err)
+	vRec.AddConfigPath(path)
+	vRec.SetConfigName("boxRecipient")
+	if err := vRec.ReadInConfig(); err != nil {
+		t.Fatalf("%s\n", err)
+	}
+
+	options, err = FromViper(vRec)
+	assert.NoError(err)
+
+	decrypter, err := options.LoadDecrypt()
+	assert.NoError(err)
+
+	msg := []byte("hello")
+	data, nonce, err := encrypter.EncryptMessage(msg)
+	assert.NoError(err)
+
+	decodedMSG, err := decrypter.DecryptMessage(data, nonce)
+	assert.NoError(err)
+
+	assert.Equal(msg, decodedMSG)
+
+}